Combine mobile and desktop wg tunnels
