+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel/wireguard/connected_tunnel.rs
@@ -2,8 +2,16 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 use std::net::{IpAddr, SocketAddr};
+#[cfg(target_os = "android")]
+use std::sync::Arc;
+#[cfg(target_os = "ios")]
+use std::time::Duration;
 
+#[cfg(target_os = "ios")]
+use dispatch2::{DispatchQueue, DispatchQueueAttr};
 use ipnetwork::IpNetwork;
+#[cfg(target_os = "ios")]
+use nym_apple_network::PathMonitor;
 use nym_authenticator_client::AuthClientMixnetListenerHandle;
 #[cfg(windows)]
 use nym_routing::{Callback, CallbackHandle, EventType};
@@ -13,15 +21,21 @@ use nym_wg_go::wireguard_go::WintunInterface;
 use nym_wg_go::{netstack, wireguard_go};
 #[cfg(windows)]
 use nym_windows::net::{self as winnet, AddressFamily};
-#[cfg(windows)]
+#[cfg(any(windows, target_os = "ios"))]
 use tokio::sync::mpsc;
 use tokio::task::{JoinError, JoinHandle};
+#[cfg(target_os = "ios")]
+use tokio_stream::{StreamExt, wrappers::UnboundedReceiverStream};
 use tokio_util::sync::CancellationToken;
 #[cfg(unix)]
 use tun::AsyncDevice;
 
+#[cfg(target_os = "android")]
+use crate::tunnel_provider::AndroidTunProvider;
 #[cfg(windows)]
 use crate::tunnel_state_machine::route_handler::RouteHandler;
+#[cfg(target_os = "ios")]
+use crate::tunnel_state_machine::tunnel::wireguard::dns64::Dns64Resolution;
 #[cfg(unix)]
 use crate::tunnel_state_machine::tunnel::wireguard::fd::DupFd;
 use crate::{
@@ -37,6 +51,10 @@ use crate::{
     },
     wg_config::{AllowedIps, WgNodeConfig},
 };
+
+/// Delay before acting on default route changes.
+#[cfg(target_os = "ios")]
+const DEFAULT_PATH_DEBOUNCE: Duration = Duration::from_millis(250);
 
 pub struct ConnectedTunnel {
     entry_gateway_client: WgGatewayClient,
@@ -78,10 +96,12 @@ impl ConnectedTunnel {
     pub async fn run(
         self,
         #[cfg(windows)] route_handler: RouteHandler,
+        #[cfg(target_os = "android")] tun_provider: Arc<dyn AndroidTunProvider>,
         options: TunnelOptions,
         tunnel_constants: TunnelConstants,
     ) -> Result<TunnelHandle> {
         match options {
+            #[cfg(not(any(target_os = "ios", target_os = "android")))]
             TunnelOptions::TunTun(tuntun_options) => {
                 self.run_using_tun_tun(
                     #[cfg(windows)]
@@ -94,12 +114,15 @@ impl ConnectedTunnel {
             TunnelOptions::Netstack(netstack_options) => self.run_using_netstack(
                 #[cfg(windows)]
                 route_handler,
+                #[cfg(target_os = "android")]
+                tun_provider,
                 netstack_options,
                 tunnel_constants,
             ),
         }
     }
 
+    #[cfg(not(any(target_os = "ios", target_os = "android")))]
     async fn run_using_tun_tun(
         self,
         #[cfg(windows)] route_handler: RouteHandler,
@@ -224,6 +247,7 @@ impl ConnectedTunnel {
     fn run_using_netstack(
         self,
         #[cfg(windows)] route_handler: RouteHandler,
+        #[cfg(target_os = "android")] tun_provider: Arc<dyn AndroidTunProvider>,
         options: NetstackTunnelOptions,
         tunnel_constants: TunnelConstants,
     ) -> Result<TunnelHandle> {
@@ -250,10 +274,31 @@ impl ConnectedTunnel {
             None,
         );
 
+        // Save entry peer so that we can re-resolve it and update wg config on network changes.
+        #[cfg(target_os = "ios")]
+        let orig_entry_peer = wg_entry_config.peer.clone();
+
         let mut two_hop_config = TwoHopConfig::new(wg_entry_config, wg_exit_config);
+
+        // iOS does not perform dns64 resolution by default. Do that manually.
+        #[cfg(target_os = "ios")]
+        two_hop_config.entry.peer.resolve_in_place()?;
 
         let mut entry_tunnel =
             netstack::Tunnel::start(two_hop_config.entry.clone().into_netstack_config())?;
+
+        // Configure tunnel sockets to bypass the tunnel interface.
+        #[cfg(target_os = "android")]
+        {
+            match entry_tunnel.get_socket_v4() {
+                Ok(fd) => tun_provider.bypass(fd),
+                Err(e) => tracing::error!("Failed to obtain bypass socket (ipv4): {}", e),
+            }
+            match entry_tunnel.get_socket_v6() {
+                Ok(fd) => tun_provider.bypass(fd),
+                Err(e) => tracing::error!("Failed to obtain bypass socket (ipv6): {}", e),
+            }
+        }
 
         // Open connection to the exit node via entry node.
         let exit_in_tunnel_udp_proxy = entry_tunnel.start_in_tunnel_udp_connection_proxy(
@@ -268,7 +313,8 @@ impl ConnectedTunnel {
             tunnel_constants.in_tunnel_bandwidth_metadata_endpoint,
         )?;
 
-        let exit_tunnel = wireguard_go::Tunnel::start(
+        #[allow(unused_mut)]
+        let mut exit_tunnel = wireguard_go::Tunnel::start(
             two_hop_config.exit.into_wireguard_config(),
             #[cfg(unix)]
             options.exit_tun.get_ref().dup_fd().map_err(Error::DupFd)?,
@@ -318,19 +364,74 @@ impl ConnectedTunnel {
                 }
             }
 
-            #[cfg(not(windows))]
+            #[cfg(not(any(windows, target_os = "ios")))]
             {
                 child_shutdown_token.cancelled().await;
                 tracing::debug!("Received tunnel shutdown event. Exiting event loop.");
             }
 
-            entry_tunnel.stop();
-            entry_magic_bandwidth_tcp_proxy.close();
-            exit_in_tunnel_udp_proxy.close();
+            #[cfg(target_os = "ios")]
+            {
+                let (default_path_tx, default_path_rx) = mpsc::unbounded_channel();
+                let mut default_path_rx = debounced::debounced(
+                    UnboundedReceiverStream::new(default_path_rx),
+                    DEFAULT_PATH_DEBOUNCE,
+                );
+
+                let queue = DispatchQueue::new(
+                    "net.nymtech.vpn.wg-path-monitor",
+                    DispatchQueueAttr::SERIAL,
+                );
+                let mut path_monitor = PathMonitor::new();
+                path_monitor.set_dispatch_queue(&queue);
+                path_monitor.set_update_handler(move |network_path| {
+                    if let Err(e) = default_path_tx.send(network_path) {
+                        tracing::error!("Failed to send new default path: {}", e);
+                    }
+                });
+                path_monitor.start();
+
+                loop {
+                    tokio::select! {
+                        Some(new_path) = default_path_rx.next() => {
+                            tracing::debug!("New default path: {}", new_path.description());
+
+                            // Depending on the network device is connected to, we may need to re-resolve the IP addresses.
+                            // For instance when device connects to IPv4-only server from IPv6-only network,
+                            // it needs to use an IPv4-mapped address, which can be received by re-resolving
+                            // the original peer IP.
+                            match orig_entry_peer.resolved() {
+                                Ok(resolved_peer) => {
+                                    let peer_update = resolved_peer.into_peer_endpoint_update();
+
+                                    // Update wireguard-go configuration with re-resolved peer endpoints.
+                                    if let Err(e) = entry_tunnel.update_peers(&[peer_update]) {
+                                       tracing::error!("Failed to update peers on network change: {}", e);
+                                    }
+                                }
+                                Err(e) => {
+                                    tracing::error!("Failed to re-resolve peer on default path update: {}", e);
+                                }
+                            }
+
+                            // Rebind wireguard-go on tun device.
+                            exit_tunnel.bump_sockets();
+                            entry_tunnel.bump_sockets();
+                        }
+                        _ = child_shutdown_token.cancelled() => {
+                            tracing::debug!("Received tunnel shutdown event. Exiting event loop.");
+                            break;
+                        }
+                    }
+                }
+            }
 
             // Windows: do not drop exit tunnel as it owns the underlying tunnel device.
             #[cfg(not(windows))]
             exit_tunnel.stop();
+            entry_tunnel.stop();
+            entry_magic_bandwidth_tcp_proxy.close();
+            exit_in_tunnel_udp_proxy.close();
 
             Tombstone {
                 tun_devices: vec![
@@ -403,6 +504,7 @@ impl ConnectedTunnel {
 
 pub enum TunnelOptions {
     /// Multihop configured using two tun adapters.
+    #[cfg(not(any(target_os = "ios", target_os = "android")))]
     TunTun(TunTunTunnelOptions),
 
     /// Multihop using single tun adapter and netstack with local UDP forwarder to wrap tunnels.
@@ -410,6 +512,7 @@ pub enum TunnelOptions {
 }
 
 /// Multihop configuration using two tun adapters.
+#[cfg(not(any(target_os = "ios", target_os = "android")))]
 pub struct TunTunTunnelOptions {
     /// Entry tunnel device.
     #[cfg(unix)]
